% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_mtx_list}
\alias{bind_mtx_list}
\title{Bind the read lines on tbl2 parameter tables into tables}
\usage{
bind_mtx_list(x, n, pnames)
}
\arguments{
\item{x}{list of vectors from read lines}

\item{n}{Number of parameters}

\item{pnames}{Vector of parameter names}
}
\value{
A tibble for each list of vectors
}
\description{
Bind the read lines on tbl2 parameter tables into tables
}
\keyword{internal}
