% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write_tbl}
\alias{write_tbl}
\title{Write SWAT+ input file which has a tabular structure.}
\usage{
write_tbl(tbl, file_path, fmt, add_lines = NULL, write_col_names = TRUE)
}
\arguments{
\item{tbl}{SWAT input table in tibble (data.frame) format.}

\item{file_path}{Write path of the SWAT+ input file.}

\item{fmt}{Character vector of format strings to define the print format of
each table column.}
}
\value{
Writes a text file table in the file path.
}
\description{
Write SWAT+ input file which has a tabular structure.
}
\keyword{internal}
